PopMovies: Start project
Starting PopMovies project to learn more about developing a full app.
The goal of the app is to present a movie browsing experience, based on popularity.
As a developer note the movie information is from themoviedb.org. It requires a developer api key to use this information. As posting this key publicly on Github is not allowed, it is stored in a seperate file that is not tracked on git. This file is in the values dir and is called apikey.xml. Inside this file is where someone wanting to run the app could place his or her own key.